wordcountaddin {wordcountaddin}	R Documentation

wordcountaddin

Description

This packages is an addin for RStudio that will count the words and characters in a plain text document. It is designed for use with R markdown documents and will exclude YAML header content, code chunks and inline code from the counts. It also computes readability statistics so you can get an idea of how easy or difficult your text is to read.

[Package wordcountaddin version 0.3.0.9000 Index]